Always apply the platform option if the engine supports it
This is breaking in that we used to support mislabeled V1 manifests
but by now any recent images should be able to support manifest V2 or
at least have the correct platform in the V1 manifest.

Legacy manifest V1 images that are identified with the wrong
platform (including out-of-date balenalib images) may fail with
'Error: failed to get destination image...' and it should be a forcing
function to use an up-to-date base image if one exists.

Change-type: major
Signed-off-by: Kyle Harding <kyle@balena.io>