Skip manifest query for intermediate build stages
This removes unnecessary queries to the registry for intermediate
stage images in the dockerfile. Querying the registry for these images
would fail in the best case or could cause false positives in the worst
case.

This commit adds an additional filter during parsing of the dockerfile
to remove any intermediate stages from the query list.

Change-type: patch