Create image contracts when creating release
This will use the contracts created from the composition (if any)
to set the `contract` property in the release images.

This requires passing a list of ImageDescriptors returned by
`compose.parse` when calling release `create`. The list is required to
avoid users creating a release without the contracts by mistake which
makes this a breaking change.

Change-type: major